model: ลบ published_date เพิ่ม follow_date + จัดไฟล์ใหม่
