+++ server/src/routes/index.ts
@@ -3,9 +3,7 @@ import express, { Response, Request } from "express";
 import auth from "./auth"
 import protect from "./protect"
 import home from './home'
-import summaries from './summaries'
-import post from './post'
-import manage from './manage'
+import posts from './posts'
 
 const router = express.Router();
 
@@ -13,11 +11,9 @@ router.get("/", (req: Request, res: Response) => {
     res.json({ message: "Welcome to the API" })
 })
 
-router.use("/auth", auth)
+router.use("/user", auth)
 router.use("/protect", protect)
-router.use('/home', home) // home อันนี้คือเอาไว้ทำอะไรนะะ
-router.use('/post', post) // ควรเป็นคำพหุนาม /posts เพราะว่าเรามองว่า post มันเป็นเรื่องของโพสต์ในระบบ แต่เราก็ไม่ได้มองว่ามันเป็นเรื่องของโพสต์ๆ ไหนเจาะจงเหมือน /user ที่เป็นการเข้าถึง user เพียง 1 เดียว
-router.use('/summaries', summaries)
-router.use('/', manage) //อันนี้อาจจะดูไม่ค่อน make sense ที่การยิงมาที่ /api แล้วเข้าไปที่ manage เพราะปกติ /api ไม่ควรเป็น API endpoint ที่ valid ในการทำงาน
+router.use('/home', home) // home อันนี้คือเอาไว้ทำอะไรนะะ >> ตอนแรกแยกเป็นหน้านึง 1 ไฟล์ อันนี้เลยเอาไว้สำหรับหน้า home เพราะมันต้องเลือก post ล่าสุดกับแนะนำมาขึ้น ตอนนี้คิดว่าลบดีกว่าเพราะมันเกี่ยวกับ post เอาไว้หน้า post ละกัน
+router.use('/posts', posts) // ควรเป็นคำพหุนาม /posts เพราะว่าเรามองว่า post มันเป็นเรื่องของโพสต์ในระบบ แต่เราก็ไม่ได้มองว่ามันเป็นเรื่องของโพสต์ๆ ไหนเจาะจงเหมือน /user ที่เป็นการเข้าถึง user เพียง 1 เดียว
 
 export default router;
